Correct getting start game forms
There was an error while getting the forms to start a game.
We get a traceback from a generic view in django :

```
 File "/home/romain/.local/share/virtualenvs/werewolf/lib/python3.8/site-packages/django/views/generic/detail.py", line 94, in get_context_data
    if self.object:
AttributeError: 'StartGame' object has no attribute 'object'
```

This happens because, `SingleObjectMixin` is not made to be used
directly. The code assume that `self.object` is set before the call of
`get_context_data`.

We use now `BaseDetailView` wich do it for use.